+++ lib/features/login/presentation/ui/login_screen.dart
@@ -1,5 +1,5 @@
-import 'package:dogdom/theme/app_icons.dart';
-import 'package:dogdom/ui/main_screen/main_screen.dart';
+import 'package:dogdom/app/theme/app_icons.dart';
+import 'package:dogdom/features/main/presentation/ui/main_screen.dart';
 import 'package:dogdom/utils/extension/context_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
