+++ lib/features/main/presentation/ui/main_screen.dart
@@ -1,4 +1,4 @@
-import 'package:dogdom/theme/app_icons.dart';
+import 'package:dogdom/app/theme/app_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
