+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:dogdom/ui/dogdom_app.dart';
+import 'package:dogdom/app/dogdom_app.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
